[loading] Moved a bunch of files so I don't have to deal with them just yet
