+++ src/loading/loading_utils.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <rx/core/string.h>
+
+namespace nova::renderer {
+} // namespace nova::renderer
