+++ src/loading/shaderpack/shaderpack_loading.cpp
