+++ src/loading/shaderpack/shaderpack_validator.cpp
